Added Activities about Update/Delete Color

+++ app/src/main/java/com/infomaniak/drive/data/cache/FileController.kt
@@ -792,9 +792,7 @@ object FileController {
             FileActivityType.FILE_TRASH -> {
                 if (returnResponse[fileId] == null || returnResponse[fileId]?.createdAt?.time == createdAt.time) { // Api fix
                     getParentFile(fileId = fileId, realm = realm)?.let { parent ->
-                        if (parent.id == currentFolder.id) {
-                            removeFile(fileId, customRealm = realm, recursive = false)
-                        }
+                        if (parent.id == currentFolder.id) removeFile(fileId, customRealm = realm, recursive = false)
                     }
                     returnResponse[fileId] = this
                 }
@@ -821,6 +819,8 @@ object FileController {
             FileActivityType.FILE_RENAME,
             FileActivityType.FILE_CATEGORIZE,
             FileActivityType.FILE_UNCATEGORIZE,
+            FileActivityType.FILE_COLOR_UPDATE,
+            FileActivityType.FILE_COLOR_DELETE,
             FileActivityType.FILE_SHARE_CREATE,
             FileActivityType.FILE_SHARE_DELETE,
             FileActivityType.FILE_SHARE_UPDATE,
